refactor(legal): extract permissions to sibling module
Split `IsCloudOrDevDeployment` and `IsOrganizationAdminOrOwner` into
`presentation/permissions.py`. Moves the `OrganizationMembership.objects.get`
lookup out of `views.py`, matches the conventional Django layout, and
keeps the membership check on the HTTP-auth layer where it belongs (it
gates the requesting user, not legal-documents domain data).
